requirements.txt
click==8.0.1
DateTime==4.3
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.1
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
psycopg2-binary==2.9.1
pytz==2021.1
SQLAlchemy==1.4.22
Werkzeug==2.0.1
zope.interface==5.4.0
